requirements.txt
keyboard==0.13.5
requests==2.20.0
selenium==3.141.0
beautifulsoup4==4.9.1
Pillow==7.2.0
